emr_serverless: reorg internals
